use copilot suggestions

+++ tt_metal/tools/profiler/kernel_profiler.hpp
@@ -292,7 +292,7 @@ inline void __attribute__((always_inline)) profiler_noc_async_write_posted(
     std::uint32_t src_local_l1_addr, std::uint64_t dst_noc_addr, std::uint32_t size, uint8_t noc = noc_index) {
     WAYPOINT("NAWW");
 #if !defined(KERNEL_BUILD)
-    constexpr uint32_t noc_mode = DM_DEDICATED_NOC;
+    constexpr uint8_t noc_mode = DM_DEDICATED_NOC;
 #endif
     DEBUG_SANITIZE_NOC_WRITE_TRANSACTION(noc, dst_noc_addr, src_local_l1_addr, size);
     ncrisc_noc_fast_write_any_len<noc_mode>(
